requirements.txt
beautifulsoup4==4.11.1
numpy~=1.22.3
pandas==1.4.2
plotly==5.8.0
scikit-learn==1.1.0
scipy==1.8.1
seaborn==0.11.2
shap==0.40.0
streamlit==1.9.0
yfinance==0.1.70
imblearn==0.0
xgboost==1.6.1
catboost==1.0.6
lightgbm==3.3.2
inflection==0.5.1
Pillow~=9.1.1
protobuf~=3.19.0
